NSInteger ALL the things! Force a 32-bit int so the nick colors are correct on 64-bit devices

+++ IRCCloud/Classes/BuffersDataSource.m
@@ -93,7 +93,7 @@
     }
 }
 
--(int)count {
+-(NSUInteger)count {
     @synchronized(_buffers) {
         return _buffers.count;
     }
