+++ IRCCloud/Classes/SettingsViewController.m
@@ -426,8 +426,8 @@
 }
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
-    int row = indexPath.row;
-    NSString *identifier = [NSString stringWithFormat:@"settingscell-%i-%i", indexPath.section, indexPath.row];
+    NSUInteger row = indexPath.row;
+    NSString *identifier = [NSString stringWithFormat:@"settingscell-%li-%li", (long)indexPath.section, (long)indexPath.row];
     UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:identifier];
     if(!cell)
         cell = [[UITableViewCell alloc] initWithStyle:UITableViewCellStyleValue1 reuseIdentifier:identifier];
